// src/app/api/model/objectsStreamValuesUpdateInner.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConditionValue } from './conditionValue';
import { ObjectsStreamValuesUpdateInnerItem } from './objectsStreamValuesUpdateInnerItem';


export interface ObjectsStreamValuesUpdateInner { 
    /**
     * <span style=\"color: #92989b; font-size: 13px\">RELEASES</span>:<span style=\"color: #d41f1c;\"> <strong>14.1 and 15.0</strong></span>\\ COV subscription ID that the update applies to
     */
    subscriptionId?: string;
    item: ObjectsStreamValuesUpdateInnerItem;
    /**
     * This property contains condition flags (if any are set) for the properties of `item`. If none of the properties have any condition flags set, then `condition` will not be included in the response or it will be empty. 
     */
    readonly condition?: { [key: string]: ConditionValue; };
    /**
     * A JSON schema that describes the properties of the `item` property 
     */
    schema?: object;
}